# Conflicts:
#	apps/core/account/mixins.py

+++ apps/core/account/mixins.py
@@ -1,5 +1,8 @@
 from django.db import transaction
-from apps.shared import ResponseController, BaseCreateMixin, BaseDestroyMixin, BaseListMixin
+
+from apps.core.company.models import Company
+from apps.core.account.models import User
+from apps.shared import ResponseController, BaseCreateMixin, BaseDestroyMixin
 from rest_framework.exceptions import ValidationError
 
 
@@ -36,6 +39,7 @@ class AccountCreateMixin(BaseCreateMixin):
         serializer = self.serializer_create(data=request.data)
         serializer.is_valid(raise_exception=True)
         instance = self.perform_create(serializer, request.user)
+        self.sync_new_user_to_map(instance, request.user.company_current_id)
         if not isinstance(instance, Exception):
             return ResponseController.created_201(self.serializer_class(instance).data)
         elif isinstance(instance, ValidationError):
@@ -48,6 +52,11 @@ class AccountCreateMixin(BaseCreateMixin):
                 instance = serializer.save(
                     tenant_current_id=user.tenant_current_id,
                 )
+                if instance.company_current_id:
+                    company_current_id = instance.company_current_id
+                    company = Company.object_normal.get(id=company_current_id)
+                    company.total_user = User.objects.filter(company_current=company_current_id).count()
+                    company.save()
             return instance
         except Exception as e:
             print(e)
@@ -64,5 +73,11 @@ class AccountDestroyMixin(BaseDestroyMixin):
 
     @staticmethod
     def perform_destroy(instance):
-        instance.delete()
+        if instance.company_current_id:
+            company_current_id = instance.company_current_id
 
+            instance.delete()
+
+            company = Company.object_normal.get(id=company_current_id)
+            company.total_user = User.objects.filter(company_current=company_current_id).count()
+            company.save()
